Update 1.9 Обновлен метод. Проведен рефакторинг

+++ ClassFor2DArray.cs
@@ -418,15 +418,13 @@ namespace Homework_4._8
       public static int[,] InputArrayInt(int[,] inputArray, int n, int m, string nameArray)
       {
          Console.WriteLine("Двумерный целочисленный массив  {0}:", nameArray);
-         double[,] outputArray = new double[n, m];
+         int[,] outputArray = new int[n, m];
          for (int i = 0; i < n; i++)
          {
             for (int j = 0; j < m; j++)
             {
                outputArray[i, j] = inputArray[i, j];
                Console.Write("{0} ", outputArray[i, j]);
-               //Console.Write("{0:f2} ", outputArray[i, j]);
-               //Console.Write("{0:f} ", outputArray[i, j]);
             }
 
             Console.WriteLine();
